fix(agent): run orphaned-settings recovery at process start from bin.ts
Settings handling moves to claude-settings.ts so bin.ts can import it
without dragging in the agent stack. Recovery now fires before yargs
dispatch — ahead of conflict detection, OAuth, and the agent — for every
command.
