+++ src/lib/agent/agent-runner.ts
@@ -29,11 +29,12 @@ import {
   AgentErrorType,
   AgentSignals,
   buildWizardMetadata,
+} from './agent-interface';
+import {
   checkAllSettingsConflicts,
   backupAndFixClaudeSettings,
   restoreClaudeSettings,
-  recoverOrphanedSettingsBackups,
-} from './agent-interface';
+} from './claude-settings';
 import { getCloudUrlFromRegion } from '@utils/urls';
 import {
   evaluateWizardReadiness,
@@ -189,12 +190,6 @@ export async function runProgram(
   if (session.debug) {
     enableDebugLogs();
   }
-
-  // Heal any settings backup an earlier interrupted run left orphaned. Must
-  // run before the conflict check below: an orphan means the conflicting
-  // settings.json is sitting in .wizard-backup, so detection against the live
-  // file would miss it and the restore would resurface the overrides mid-run.
-  recoverOrphanedSettingsBackups(session.installDir);
 
   const skillsBaseUrl = getSkillsBaseUrl(session.localMcp);
